# cardano-rosetta-server/src/server/openApi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.4.1",
    "title": "Rosetta",
    "description": "Build Once. Integrate Your Blockchain Everywhere.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/construction/derive": {
      "post": {
        "summary": "Derive an Address from a PublicKey",
        "description": "Derive returns the network-specific address associated with a public key. Blockchains that require an on-chain action to create an account should not implement this method.",
        "operationId": "constructionDerive",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionDeriveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionDeriveResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/preprocess": {
      "post": {
        "summary": "Create a Request to Fetch Metadata",
        "description": "Preprocess is called prior to `/construction/payloads` to construct a request for any metadata that is needed for transaction construction given (i.e. account nonce). The request returned from this method will be used by the caller (in a different execution environment) to call the `/construction/metadata` endpoint.",
        "operationId": "constructionPreprocess",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionPreprocessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionPreprocessResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/metadata": {
      "post": {
        "summary": "Get Metadata for Transaction Construction",
        "description": "Get any information required to construct a transaction for a specific network. Metadata returned here could be a recent hash to use, an account sequence number, or even arbitrary chain state. The request used when calling this endpoint is often created by calling `/construction/preprocess` in an offline environment. It is important to clarify that this endpoint should not pre-construct any transactions for the client (this should happen in `/construction/payloads`). This endpoint is left purposely unstructured because of the wide scope of metadata that could be required.",
        "operationId": "constructionMetadata",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionMetadataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionMetadataResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/payloads": {
      "post": {
        "summary": "Generate an Unsigned Transaction and Signing Payloads",
        "description": "Payloads is called with an array of operations and the response from `/construction/metadata`. It returns an unsigned transaction blob and a collection of payloads that must be signed by particular addresses using a certain SignatureType. The array of operations provided in transaction construction often times can not specify all \"effects\" of a transaction (consider invoked transactions in Ethereum). However, they can deterministically specify the \"intent\" of the transaction, which is sufficient for construction. For this reason, parsing the corresponding transaction in the Data API (when it lands on chain) will contain a superset of whatever operations were provided during construction.",
        "operationId": "constructionPayloads",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionPayloadsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionPayloadsResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/combine": {
      "post": {
        "summary": "Create Network Transaction from Signatures",
        "description": "Combine creates a network-specific transaction from an unsigned transaction and an array of provided signatures. The signed transaction returned from this method will be sent to the `/construction/submit` endpoint by the caller.",
        "operationId": "constructionCombine",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionCombineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionCombineResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/parse": {
      "post": {
        "summary": "Parse a Transaction",
        "description": "Parse is called on both unsigned and signed transactions to understand the intent of the formulated transaction. This is run as a sanity check before signing (after `/construction/payloads`) and before broadcast (after `/construction/combine`). ",
        "operationId": "constructionParse",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionParseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstructionParseResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/hash": {
      "post": {
        "summary": "Get the Hash of a Signed Transaction",
        "description": "TransactionHash returns the network-specific transaction hash for a signed transaction.",
        "operationId": "constructionHash",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionHashRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionIdentifierResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/construction/submit": {
      "post": {
        "summary": "Submit a Signed Transaction",
        "description": "Submit a pre-signed transaction to the node. This call should not block on the transaction being included in a block. Rather, it should return immediately with an indication of whether or not the transaction was included in the mempool. The transaction submission response should only return a 200 status if the submitted transaction could be included in the mempool. Otherwise, it should return an error.",
        "operationId": "constructionSubmit",
        "tags": ["Construction"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstructionSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionIdentifierResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NetworkIdentifier": {
        "description": "The network_identifier specifies which network a particular object is associated with.",
        "type": "object",
        "required": ["blockchain", "network"],
        "properties": {
          "blockchain": {
            "type": "string",
            "example": "bitcoin"
          },
          "network": {
            "description": "If a blockchain has a specific chain-id or network identifier, it should go in this field. It is up to the client to determine which network-specific identifier is mainnet or testnet.",
            "type": "string",
            "example": "mainnet"
          },
          "sub_network_identifier": {
            "$ref": "#/components/schemas/SubNetworkIdentifier"
          }
        }
      },
...
      "PublicKey": {
        "description": "PublicKey contains a public key byte array for a particular CurveType encoded in hex. Note that there is no PrivateKey struct as this is NEVER the concern of an implementation.",
        "type": "object",
        "required": ["hex_bytes", "curve_type"],
        "properties": {
          "hex_bytes": {
            "type": "string",
            "description": "Hex-encoded public key bytes in the format specified by the CurveType."
          },
          "curve_type": {
            "$ref": "#/components/schemas/CurveType"
          }
        }
      },
...
      "SigningPayload": {
        "description": "SigningPayload is signed by the client with the keypair associated with an address using the specified SignatureType. SignatureType can be optionally populated if there is a restriction on the signature scheme that can be used to sign the payload.",
        "type": "object",
        "required": ["address", "hex_bytes"],
        "properties": {
          "address": {
            "type": "string",
            "description": "The network-specific address of the account that should sign the payload."
          },
          "hex_bytes": {
            "type": "string"
          },
          "signature_type": {
            "$ref": "#/components/schemas/SignatureType"
          }
        }
      },
...
      "ConstructionMetadataRequest": {
        "description": "A ConstructionMetadataRequest is utilized to get information required to construct a transaction. The Options object used to specify which metadata to return is left purposely unstructured to allow flexibility for implementers.",
        "type": "object",
        "required": ["network_identifier", "options"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "options": {
            "description": "Some blockchains require different metadata for different types of transaction construction (ex: delegation versus a transfer). Instead of requiring a blockchain node to return all possible types of metadata for construction (which may require multiple node fetches), the client can populate an options object to limit the metadata returned to only the subset required.",
            "type": "object",
            "required": ["relative_ttl"],
            "properties": {
              "relative_ttl": { "type": "number" }
            }
          }
        }
      },
      "ConstructionMetadataResponse": {
        "description": "The ConstructionMetadataResponse returns network-specific metadata used for transaction construction.",
        "type": "object",
        "required": ["metadata"],
        "properties": {
          "metadata": {
            "type": "object",
            "properties": {
              "ttl": { "type": "string" }
            },
            "example": {
              "account_sequence": 23,
              "recent_block_hash": "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
            }
          }
        }
      },
      "ConstructionDeriveRequest": {
        "description": "ConstructionDeriveRequest is passed to the `/construction/derive` endpoint. Network is provided in the request because some blockchains have different address formats for different networks. Metadata is provided in the request because some blockchains allow for multiple address types (i.e. different address for validators vs normal accounts).",
        "type": "object",
        "required": ["network_identifier", "public_key"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "public_key": {
            "$ref": "#/components/schemas/PublicKey"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "ConstructionDeriveResponse": {
        "description": "ConstructionDeriveResponse is returned by the `/construction/derive` endpoint.",
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "type": "string",
            "description": "Address in network-specific format."
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "ConstructionPreprocessRequest": {
        "description": "ConstructionPreprocessRequest is passed to the `/construction/preprocess` endpoint so that a Rosetta implementation can determine which metadata it needs to request for construction.",
        "type": "object",
        "required": ["network_identifier", "operations"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "metadata": {
            "type": "object",
            "required": ["relative_ttl"],
            "properties": {
              "relative_ttl": { "type": "number" }
            }
          }
        }
      },
      "ConstructionPreprocessResponse": {
        "description": "ConstructionPreprocessResponse contains the request that will be sent directly to `/construction/metadata`. If it is not necessary to make a request to `/construction/metadata`, options should be null.",
        "type": "object",
        "properties": {
          "options": {
            "type": "object",
            "description": "The options that will be sent directly to `/construction/metadata` by the caller.",
            "required": ["relative_ttl"],
            "properties": {
              "relative_ttl": { "type": "number" }
            }
          }
        }
      },
      "ConstructionPayloadsRequest": {
        "description": "ConstructionPayloadsRequest is the request to `/construction/payloads`. It contains the network, a slice of operations, and arbitrary metadata that was returned by the call to `/construction/metadata`.",
        "type": "object",
        "required": ["network_identifier", "operations", "metadata"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "metadata": {
            "type": "object",
            "required": ["ttl"],
            "properties": {
              "ttl": { "type": "string" }
            }
          }
        }
      },
      "ConstructionPayloadsResponse": {
        "description": "ConstructionTransactionResponse is returned by `/construction/payloads`. It contains an unsigned transaction blob (that is usually needed to construct the a network transaction from a collection of signatures) and an array of payloads that must be signed by the caller.",
        "type": "object",
        "required": ["unsigned_transaction", "payloads"],
        "properties": {
          "unsigned_transaction": {
            "type": "string"
          },
          "payloads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SigningPayload"
            }
          }
        }
      },
      "ConstructionCombineRequest": {
        "description": "ConstructionCombineRequest is the input to the `/construction/combine` endpoint. It contains the unsigned transaction blob returned by `/construction/payloads` and all required signatures to create a network transaction.",
        "type": "object",
        "required": ["network_identifier", "unsigned_transaction", "signatures"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "unsigned_transaction": {
            "type": "string"
          },
          "signatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Signature"
            }
          }
        }
      },
      "ConstructionCombineResponse": {
        "description": "ConstructionCombineResponse is returned by `/construction/combine`. The network payload will be sent directly to the `construction/submit` endpoint.",
        "type": "object",
        "required": ["signed_transaction"],
        "properties": {
          "signed_transaction": {
            "type": "string"
          }
        }
      },
      "ConstructionParseRequest": {
        "description": "ConstructionParseRequest is the input to the `/construction/parse` endpoint. It allows the caller to parse either an unsigned or signed transaction.",
        "type": "object",
        "required": ["network_identifier", "signed", "transaction"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "signed": {
            "type": "boolean",
            "description": "Signed is a boolean indicating whether the transaction is signed."
          },
          "transaction": {
            "type": "string",
            "description": "This must be either the unsigned transaction blob returned by `/construction/payloads` or the signed transaction blob returned by `/construction/combine`."
          }
        }
      },
      "ConstructionParseResponse": {
        "description": "ConstructionParseResponse contains an array of operations that occur in a transaction blob. This should match the array of operations provided to `/construction/preprocess` and `/construction/payloads`.",
        "type": "object",
        "required": ["operations", "signers"],
        "properties": {
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "signers": {
            "description": "All signers of a particular transaction. If the transaction is unsigned, it should be empty.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "ConstructionHashRequest": {
        "description": "ConstructionHashRequest is the input to the `/construction/hash` endpoint.",
        "type": "object",
        "required": ["network_identifier", "signed_transaction"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "signed_transaction": {
            "type": "string"
          }
        }
      },
      "ConstructionSubmitRequest": {
        "description": "The transaction submission request includes a signed transaction.",
        "type": "object",
        "required": ["network_identifier", "signed_transaction"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "signed_transaction": {
            "type": "string"
          }
        }
      },
      "TransactionIdentifierResponse": {
        "description": "TransactionIdentifierResponse contains the transaction_identifier of a transaction that was submitted to either `/construction/hash` or `/construction/submit`.",
        "type": "object",
        "required": ["transaction_identifier"],
        "properties": {
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "Error": {
        "description": "Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.",
        "type": "object",
        "required": ["code", "message", "retriable"],
        "properties": {
          "code": {
            "description": "Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 12
          },
          "message": {
            "description": "Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.",
            "type": "string",
            "example": "Invalid account format"
          },
          "retriable": {
            "description": "An error is retriable if the same request may succeed if submitted again.",
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" }
            },
            "description": "Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.",
            "example": {
              "address": "0x1dcc4de8dec75d7aab85b567b6",
              "error": "not base64"
            }
          }
        }
      }
    }
  }
}